+++ Prolangle/Snippets/Brainfuck/TerseBrainfuckSnippet.cs
@@ -5,7 +5,7 @@ namespace Prolangle.Snippets.Brainfuck;
 /// <summary>
 /// https://en.wikipedia.org/w/index.php?title=Brainfuck&oldid=1216286405#Hello_World!
 /// </summary>
-public class TerseBrainfuckSnippet : ICodeSnippet
+public class TerseBrainfuckSnippet : IAttributedCodeSnippet, ILicensedCodeSnippet
 {
 	public ILanguage Language => Languages.Brainfuck.Instance;
 
@@ -13,4 +13,7 @@ public class TerseBrainfuckSnippet : ICodeSnippet
 
 	public string SourceCode =>
 		"++++++++[>++++[>++>+++>+++>+<<<<-]>+>+>->>+[<]<-]>>.>---.+++++++..+++.>>.<-.<.+++.------.--------.>>+.>++.";
+
+	public string Attribution => "https://en.wikipedia.org/w/index.php?title=Brainfuck&oldid=1216286405#Hello_World!";
+	public string License => "Creative Commons Attribution-ShareAlike License";
 }