+++ Prolangle/Snippets/IAttributedCodeSnippet.cs
@@ -1,0 +1,10 @@
+namespace Prolangle.Snippets;
+
+public interface IAttributedCodeSnippet : ICodeSnippet
+{
+	/// <summary>
+	/// The attribution for this code snippet. Will be revealed when the user
+	/// has won.
+	/// </summary>
+	string Attribution { get; }
+}